# Optimize QueryBuilder.__copy__ some more