@grigi 2019-01-17
@twheys I fixed an oversight, it should be correct now.
Please review.